requirements.txt
PyYAML==3.11
argparse==1.2.1
elasticsearch==1.0.0
psutil==2.1.1
requests==2.3.0
urllib3==1.8.3
wsgiref==0.1.2
